API

MDPAgent

AlgebraicRL.MDPAgent — Type

mutable struct MDPAgent
    mdp::AbstractEnv
    agent::AbstractPolicy
end

Defines a discrete machine that consists of an MDP and the agent capable of solving it. Running solve() will iterate through 1 episode of the MDP and return a done signal. The MDP follows the specification from ReinforcementLearning.jl. Each MDP, of type AbstractEnv, has the following functions:

action_space(env::YourEnv)

state(env::YourEnv)

state_space(env::YourEnv)

reward(env::YourEnv)

is_terminated(env::YourEnv)

reset!(env::YourEnv)

(env::YourEnv)(action) # note this function applies your action to the environment and steps forward in time.

See https://juliareinforcementlearning.org/docs/Howtowriteacustomized_environment/ for more details on how to write a custom environment. Alternatively, see https://github.com/JuliaReinforcementLearning/ReinforcementLearning.jl/tree/master/src/ReinforcementLearningExperiments/deps/experiments/experiments for examples of used environments.

The agent also comes from ReinforcementLearning.jl. See https://github.com/JuliaReinforcementLearning/ReinforcementLearning.jl/blob/61256bcf1c493914d5003f22e126c997332c2c39/src/ReinforcementLearningBase/src/interface.jl for a full definition. Mainly, it defines the following functions:

(policy::AbstractPolicy)(env) = returns an action from the policy given the environment

update(policy::AbstractPolicy, experience) = updates the policy based on an online or offline experience

AlgebraicRL.solve — Function

solve(mdpAgent::MDPAgent, display = false)

Runs 1 episode of the environment under the policy provided by the agent.

Arguments:

mdpAgent - The MDPAgent you constructed, which contains an MDP following the AbstractEnv interface and an agent following the AbstractPolicy interface.

display - Whether or not to display the environment each step. Displaying is done via calling Base.display, so overwrite Base.display(yourEnv) in order to view your environment being solved.

Returns:

The accumulated reward over the episode as returned by the environment

Base.collect — Method

Base.collect(mdpagent::MDPAgent)

Overwrites base.collect for our class since it will be used by Algebraic dynamics. Simply returns the mdpAgent since there is nothing to collect. Note collect is usually used to return an array of elements from a collection or iterator

MDPAgentMachine

AlgebraicRL.MDPAgentMachine — Method

MDPAgentMachine(    env::AbstractEnv, agent::AbstractPolicy;
                    input_size = 1, input_function = default_input_function,
                    output_size = 1, output_function = default_output_function,
                    display = false, resetMDP = true
                    )

This function returns a DiscreteMachine from AlgebraicDynamics

The env and agent must follow the respective interfaces from ReinforcementLearning.jl

By default, the internal workings of the machine does not use the input provided to eval_dynamics. This can be overwritten by using a custom input function. Make sure the size of the vector in your input function is the same as the input_size you provide.

Its output is true if the MDP is terminated and false otherwise. This can also be overwritten by providing a custom output function. Once again, make sure the size of the output vector is the same as the output_size you provide in this function call.

You can also reset the environment before running the machine if you wish. This is useful in some cases to make sure the environment is in a good state, but it is bad in other cases. So, you are able to control if it is reset or not.

You can also choose to display the environment each timestep by setting display = true. It will call Base.display(env), so you must overwrite Base.display for your environment in order to use that functionality.

Arguments:

env - a MDP following the AbstractEnv interface

agent - a agent/policy following the AbstractPolicy interface

input_size - an int which must be equal to the size of the vector used by the input function

input_function - a function which is called on the MDP before running an episode. Must have the same arguments as default_input_function

output_size - an int which must be equal to the size of the vector returned by the output function

output_function - a function which is called after the episode is terminated. Must have the same arguments as default_input_function. Returns a vector.

display - whether or not to display the MDP every timestep. Calls Base.display(env) if true.

resetMDP - whether or not to reset the MDP via reset!(env) before running an episode.

AlgebraicRL.default_output_function — Method

default_output_function(env::AbstractEnv)

Returns the result of the mdp episode. Is it done or not? To do a more sophisticated readout, such as reading internal state, write your own function with the same arguments that returns a vector. Output vector length must be the same length as specified by output_size in the function call to MDPAgentMachine

Arguments:

env - A MDP following the AbstractEnv interface

Returns:

1.0 if the environment is terminated, 0.0 otherwise.

AlgebraicRL.default_input_function — Method

default_input_function(env::AbstractEnv, internal_state::Vector)

Sets the internal state of the mdp to be the internal state provided. Default function does nothing. To customize, create your own function with the same arguments. The internal state must be the same size as the size given by input_size in the function call to MDPAgentMachine

Arguments:

env - A MDP following the AbstractEnv interface

internal_state - A vector which contains the information needed to set the internal state. In the default_input_function, does nothing.

Returns:

nothing

MDPAgentMachine and AlgebraicDynamics.jl

The following relates to MDPAgentMachines and how they interact with the AlgebraicDynamics.jl interface. We are not using internal variables, which they call 'u', so we are able to simplify the function calls slightly.

AlgebraicDynamics.UWDDynam.eval_dynamics — Function

AlgebraicDynamics.DWDDynam.eval_dynamics(f::AbstractMachine, xs=[0.0])

Runs an episode of the MDP following the policy given in the constructor for the machine. Is equilvalent to calling eval_dynamics with zeros for the internal state variable.

Arguments:

f - a machine created by MDPAgentMachine

xs - external variables used for setting the initial state

Returns:

a vector of zeros. IE basically nothing but AlgebraicDynamics requires a vector.

AlgebraicRL.read_output — Method

read_output(f::AbstractMachine)

Since we are not using internal variables via AlgebraicDynamics, we can overwrite readout to assume zeros for those variables. This is convenient for our purposes. However, due to their API structure, we would have issues if we tried to overwite readout(::AbstractMachine). Instead, we will simply rename the function to read_output for our purposes.

Arguments:

f - a machine created by MDPAgentMachine

Returns:

The result of the output_function specified in the function call to MDPAgentMachine